Added attribute command Added bar command

+++ c/motor/mmtest/src/mmmotor_settings.h
@@ -13,7 +13,8 @@
 #ifndef MMMOTOR_SETTINGS_H
 #define	MMMOTOR_SETTINGS_H
 
-typedef int16_t CHANNEL_VAL;
+//typedef int16_t CHANNEL_VAL;
+typedef int32_t CHANNEL_VAL;
 
 #define CHANNELS    8      // nr of channels
 #define ALARMS      10      // nr of alarms
